Separate classes in single-class file #14
